current index didn't change after using the 'runs_on_provider', move function 'check_default_cluster_context_index_equal_to_current_index' to the file 'ocs_ci/ocs/managedservice.py'
